requirements.txt
discord.py==2.0.0
requests==2.28.1
